+++ gen-rs/src/modeling/dists/mvnormal.rs
@@ -1,16 +1,16 @@
 use rand::rngs::ThreadRng;
 use super::{Distribution,normal};
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 use nalgebra::{DVector,DMatrix};
 
 
 pub struct MvNormal { }
 pub const mvnormal: MvNormal = MvNormal { };
 
-impl Distribution<DVector<f32>,(DVector<f32>,DMatrix<f32>)> for MvNormal {
-    fn logpdf(&self, x: &DVector<f32>, params: &(DVector<f32>,DMatrix<f32>)) -> f32 {
+impl Distribution<DVector<f64>,(DVector<f64>,DMatrix<f64>)> for MvNormal {
+    fn logpdf(&self, x: &DVector<f64>, params: &(DVector<f64>,DMatrix<f64>)) -> f64 {
         let (mu, cov) = params;
-        let k = mu.len() as f32;
+        let k = mu.len() as f64;
         let cov_det = cov.determinant();
         let cov_inv = cov.clone().try_inverse().unwrap();
         let centered_x = x - mu;
@@ -18,9 +18,9 @@ impl Distribution<DVector<f32>,(DVector<f32>,DMatrix<f32>)> for MvNormal {
         -(k*(2.*PI).ln() + cov_det.ln() + mahalanobis_squared)/2.
     }
 
-    fn random(&self, rng: &mut ThreadRng, params: &(DVector<f32>,DMatrix<f32>)) -> DVector<f32> {
+    fn random(&self, rng: &mut ThreadRng, params: &(DVector<f64>,DMatrix<f64>)) -> DVector<f64> {
         let (mu, cov) = params;
-        let transform: DMatrix<f32>;
+        let transform: DMatrix<f64>;
         let decomp = cov.clone().symmetric_eigen();
         match cov.clone().cholesky() {
             Some(c) => {
